Cleaning Code and Almost working Input, se Description for more info
It seems that both hands are working, but when i use the other controller to pick up something and the other hand is holding a item. the other hand will drop the item.

Needs more testing, not sure why. But now you could atleast pick up and throw one item.